requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
blinker==1.9.0
boto3==1.40.39
botocore==1.40.39
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
flask-swagger-ui==5.21.0
flask-talisman==1.1.0
gevent==25.9.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==6.0
zope.interface==8.0.1
setuptools>=78.1.1 # not directly required, pinned by Snyk to avoid a vulnerability